first commit
I've included nearly everything that I think would be good to have in
this package. This does not mean that I've heavily tested or optimized
it (although much of it has gone through this in the packages I'm
transferring it from).
TODO:
* Project compats: not in place because moving some stuff here will
  result in breaking changes on depend packages (mostly AxisIndices)
* Documentation: Many docstrings are already in place. Some are messy
  and it would be good to add examples to more as things are tested.
* Tests/Optimization: These need to be transferred to this repo still. I
  haven't done so already because They are currently unorganized and
  messy.

The plan is to go through the docs/tests/performance optimization in
following order for subsequent PRs:
1. "names.jl"
2. "times.jl"
3. "spatial.jl"
4. "iterators.jl"