Figured out how to isolate spatial names
One of the issues with spatial dimensions is that there can be multiple
and they have many possible names. The JuliaImages ecosystem approached
this by considering the first 3 dimensions that weren't attributed to
time or color as spatial. The easiest way of managing this is to
define all dimensions that aren't spatial and then have a method that
checks for the presence of these other names before calling something a
spatial dimension. However, this means that only those names originally
defined as part of the this package could really work with spatial
dims, so people couldn't define their own dimensions and specify it
wasn't spatial. This meant I was trying to create a minimally inclusive
list of names that definitevly weren't spatial and others might be
useful to others (obviously a bad approach).

This new commit gets away from that by keeping dimension names
propagated through the `spatialdims` sub-methods as a type parameter
into a @generated function. A careful pipeline, establish what is
spatial pass it on to another generated function that drops non spatial,
etc. Now users can define a dimension as not spatil using
`SpatioTemporalTraits.@not_spatial :not_spatial_dimname`. Even better,
users can now call `@defdim Name, name, error_method_if_not_found` to
define a new type of semantic dimension. This also creates a new macro
that has the naming pattern `@is_<name> :some_name`, which allows users
to have a universal `<name>dim` method for extracting anything that
corresponds to a dimension that is synonymous with "name". `@is_<name>`
also calls `@not_spatial` under the hood so that users don't have to
worry about that.