Take advantage of developments in ArrayInterface
This is more true to being a "Traits" package now as it is simplified to
only a single dependency and impliments minimal methods. There's still
the option to bring back or create more methods in the future. However,
this will hopefully help encourage adoption of this package as it is low
maintenance and easy to build on now.