Use ArrayInterface.jl's to_dims and Static.StaticSymbol
This moves away from the big macro that was at the center of a lot of
this before.
* Using `StaticSymbol` eliminates the need to cater to constant
propagation of a `Symbol` at every step.
* Using `ArrayInterface.to_dims` eliminates the need to define custom
`eachslice` and other dimension related methods for named dimensions so
why bother generating a bunch of custom functions.
* It's still relatively easy to make a new name work for the time
dimension or state that it is not a spatial dimension, but I want to
think about making that less public b/c the vast majority of the time
people don't need to be redefining this. I also worry about it causing
invalidations if people start using it incorrectly.